Setup javascript testing
The package.json file is used as our task runner. It runs the
server so that we can launch the specrunner.html, and watches
for changes in the javascript files and the views.

While testing the javascript files, changes in the javascript
folders will trigger the linter, and changes in the views will
trigger the ruby tests. This is to make sure that we don't break
the ruby tests when we change things in the frontend.

To test the frontend do:

    npm test

then load localhost:9292/javascripts/specrunner.html