Add a wrapper for the Stripe JS library
This wrapper aims to isolate the application from the Stripe
library. The implementation follows the documentation for
the custom integration in the Detailed Checkout Guide page:

    https://stripe.com/docs/checkout#integration-custom

Before opening the Checkout pop up, the amount is checked.
If the validation passes, the pop up launches. If not, it
defaults to the browser native validation, which won't
allow you to submit the form until the field is filled,
and in modern browsers, until the field is filled and is
a number.