Add a presenter for the success view
This presenter separates the content from the logic behind it and
allows to keep logic out of the views. Also, separating the strings
from the view may help in the future with translations, although
now it is too early to think about that, so for the time being this
is done just for separation of concerns.