Plug the presenter for the success page in the routes
For the time being I am using our current CDN at Amazon, since
we haven't decided yet if we try a new one or stick with Amazon.

I also introduced Share This for the sharing buttons, since I
know of them for a long time now and they have support for
wassap share, which the Spanish office wanted. Share This also
does some analytics on the clicks that could be useful for
campaigners or anybody looking at that kind of stats.