Add a class to deal with appending a timestamp
The reason to append a timestamp to styles and JavaScripts is to
force browsers to recache these files when they have changed in
staging or in production.

This class also generates both the local and upstream paths to
the datestamped files, so that they can be used in other deploy
files.