Add a class to deal with appending a digest
The reason to append a digest to styles and JavaScripts is to
force browsers to recache these files when they have changed in
staging or in production.

This class also generates both the local and upstream paths to
the digested files, so that they can be used by other deploy
classes.